Introducing unused_deps_py
This is a re-write of unused_deps in Python with added support for
maven dependencies from jvm_import from rules_jvm_external, which is
the default provider for maven_install dependencies.

This tool is smart enough to work with multiple maven_install
repositories and if you have both pinned and unpinned it will only use
the pinned one to speed up things.